requirements.txt
opencv-python==4.6.0.66
keyboard==0.13.5
numpy==1.24.0
pyautogui==0.9.53
pywin32==305
Pillow==9.3.0